Add tests to verify API V3 behavior on JobTemplate#launch
https://bugzilla.redhat.com/show_bug.cgi?id=1369842
